fix: CodeRabbit 제안 반영 + 스크립트 역할 재분배
