+++ bootstrap/shoppingmall_seed.py
@@ -11,22 +11,37 @@
 2) 기본 쇼핑몰 데이터 시드
 3) 백오피스 데이터 시드
 4) 초기 리뷰 샘플(30건) 시드
+
+기본 동작:
+- 독립 실행 재진입 가능성을 위해, --rebuild-schema가 없더라도
+  기존 shop_* 데이터를 먼저 비운 뒤 시드한다.
 """
 
 from __future__ import annotations
 
 import argparse
 import json
+import re
 import sys
 from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
 
 from _bootstrap_common import (  # type: ignore[import-not-found]
+    detect_database_url,
+    ensure_database_exists,
+    ensure_postgres_running,
+    ensure_tools,
     error,
     info,
+    parse_database_url,
+    print_table_summary,
+    psql_query,
+    run_command,
     set_log_prefix,
+    table_exists,
 )
+from sqlalchemy import inspect, text
 
 # 실행 위치와 무관하게 shopping_mall/backend를 import 루트로 맞춘다.
 ROOT = Path(__file__).resolve().parents[1]
@@ -77,6 +92,30 @@ EXPECTED_ROW_COUNTS = {
     "shop_chat_logs": 5,
 }
 LOG_PREFIX = "S.Mall-S"
+SHOP_TABLES = [
+    "shop_categories",
+    "shop_stores",
+    "shop_products",
+    "shop_users",
+    "shop_cart_items",
+    "shop_orders",
+    "shop_order_items",
+    "shop_reviews",
+    "shop_wishlists",
+    "shop_shipments",
+    "shop_harvest_schedules",
+    "shop_revenue_entries",
+    "shop_expense_entries",
+    "shop_weekly_reports",
+    "shop_customer_segments",
+    "shop_chat_logs",
+    "shop_chat_sessions",
+]
+READY_ROW_COUNTS = {
+    **EXPECTED_ROW_COUNTS,
+    "shop_reviews": 1000,
+    "shop_chat_sessions": 0,
+}
 
 
 # shop_categories 구성 정보
@@ -840,40 +879,213 @@ def print_summary(db) -> None:
         print(f"  - {table}: {actual}{suffix}")
 
 
+def _quote_identifier(name: str) -> str:
+    return '"' + name.replace('"', '""') + '"'
+
+
+def truncate_existing_shop_data() -> None:
+    """현재 존재하는 shop_* 관리 테이블 데이터를 비운다."""
+    managed_tables = [
+        table_name
+        for table_name in sorted(Base.metadata.tables.keys())
+        if table_name.startswith("shop_")
+    ]
+    existing_tables = set(inspect(engine).get_table_names())
+    targets = [table for table in managed_tables if table in existing_tables]
+    if not targets:
+        info("truncate 대상 shop_* 테이블이 없습니다.")
+        return
+
+    info("ShoppingMall 기존 데이터 비우기(truncate)")
+    quoted_targets = ", ".join(_quote_identifier(table) for table in targets)
+    with engine.begin() as conn:
+        conn.execute(text(f"TRUNCATE TABLE {quoted_targets} RESTART IDENTITY CASCADE;"))
+
+
+def _to_sync_db_url(raw_db_url: str) -> str:
+    url = re.sub(r"^postgresql\+\w+://", "postgresql+psycopg2://", raw_db_url, count=1)
+    if url.startswith("postgresql://"):
+        return url.replace("postgresql://", "postgresql+psycopg2://", 1)
+    return url
+
+
+def uv_sync_backend(skip_sync: bool) -> None:
+    if skip_sync:
+        info("uv sync 생략 (--skip-sync)")
+        return
+    info("shopping_mall/backend 의존성 동기화(uv sync) - 시간이 많이 걸릴 수 있습니다")
+    run_command(["uv", "sync"], cwd=SHOP_BACKEND_DIR)
+
+
+def run_review_seed_pipeline(raw_db_url: str) -> None:
+    info("쇼핑몰 리뷰 시드 스크립트 실행")
+    review_seed = ROOT / "bootstrap" / "shoppingmall_review_seed.py"
+    run_command(
+        ["uv", "run", "python", str(review_seed)],
+        cwd=SHOP_BACKEND_DIR,
+        env_overrides={"DATABASE_URL": _to_sync_db_url(raw_db_url)},
+    )
+
+
+def run_seed_pipeline(raw_db_url: str, rebuild_schema: bool) -> None:
+    info("쇼핑몰 코어 시드 실행")
+    seed_script = ROOT / "bootstrap" / "shoppingmall_seed.py"
+    command = ["uv", "run", "python", str(seed_script), "--mode", "seed"]
+    if rebuild_schema:
+        command.append("--rebuild-schema")
+    else:
+        command.append("--preserve-existing-data")
+    run_command(
+        command,
+        cwd=SHOP_BACKEND_DIR,
+        env_overrides={"DATABASE_URL": _to_sync_db_url(raw_db_url)},
+    )
+
+
+def all_shop_tables_exist(db_conf: dict[str, str]) -> bool:
+    return all(table_exists(db_conf, table) for table in SHOP_TABLES)
+
+
+def truncate_shop_tables(db_conf: dict[str, str]) -> None:
+    info("기존 shop_* 데이터 비우기(truncate)")
+    targets = ", ".join(_quote_identifier(table) for table in SHOP_TABLES)
+    truncate_sql = (
+        "BEGIN; "
+        "SET LOCAL lock_timeout = '5s'; "
+        f"TRUNCATE TABLE {targets} RESTART IDENTITY CASCADE; "
+        "COMMIT;"
+    )
+    psql_query(db_conf, truncate_sql)
+
+
+def is_shoppingmall_ready(db_conf: dict[str, str]) -> bool:
+    for table, expected in READY_ROW_COUNTS.items():
+        if not table_exists(db_conf, table):
+            return False
+        actual = int(psql_query(db_conf, f"SELECT COUNT(*) FROM {table};") or "0")
+        if actual < expected:
+            return False
+    return True
+
+
+def print_db_summary(db_conf: dict[str, str], verbose_table_info: bool) -> None:
+    print_table_summary(
+        db_conf,
+        "ShoppingMall",
+        SHOP_TABLES,
+        verbose_table_info=verbose_table_info,
+    )
+
+
 def main() -> int:
     parser = argparse.ArgumentParser(description="ShoppingMall 시드 스크립트")
+    parser.add_argument("--database-url", help="DATABASE_URL 강제 지정")
+    parser.add_argument("--skip-sync", action="store_true", help="uv sync 생략")
+    parser.add_argument(
+        "--mode",
+        choices=("seed", "init", "ensure"),
+        default="init",
+        help="seed=시드만 실행, init=항상 초기화, ensure=필요 시 초기화",
+    )
     parser.add_argument(
         "--rebuild-schema",
         action="store_true",
         help="shop_* 스키마를 drop/create로 재생성한 뒤 시드합니다.",
     )
+    parser.add_argument(
+        "--preserve-existing-data",
+        action="store_true",
+        help="기존 데이터를 유지하고 시드를 추가합니다(중복 키 충돌 가능).",
+    )
+    parser.add_argument(
+        "--verbose-table-info",
+        action="store_true",
+        help="DB 요약에 테이블 컬럼/row 수 상세 정보를 출력합니다.",
+    )
     args = parser.parse_args()
 
-    set_log_prefix(LOG_PREFIX)
-    if args.rebuild_schema:
-        info("ShoppingMall 스키마 재생성(drop/create) 시작")
-        Base.metadata.drop_all(bind=engine)
-        Base.metadata.create_all(bind=engine)
-    else:
-        info("ShoppingMall 스키마 유지(create_all) 모드 시작")
-        Base.metadata.create_all(bind=engine)
-
-    db = SessionLocal()
     try:
-        state = seed_core_data(db)
-        seed_backoffice_data(db, state)
-        db.commit()
-        print()
-        print_summary(db)
-        print()
-        info("ShoppingMall 시드 완료")
+        set_log_prefix(LOG_PREFIX)
+        if args.mode == "seed":
+            if args.rebuild_schema:
+                info("ShoppingMall 스키마 재생성(drop/create) 시작")
+                Base.metadata.drop_all(bind=engine)
+                Base.metadata.create_all(bind=engine)
+            else:
+                info("ShoppingMall 스키마 유지(create_all) 모드 시작")
+                Base.metadata.create_all(bind=engine)
+
+            if args.preserve_existing_data:
+                info(
+                    "기존 데이터 유지 모드(--preserve-existing-data): 중복 키 충돌 가능"
+                )
+            else:
+                truncate_existing_shop_data()
+
+            db = SessionLocal()
+            try:
+                state = seed_core_data(db)
+                seed_backoffice_data(db, state)
+                db.commit()
+                print()
+                print_summary(db)
+                print()
+                info("ShoppingMall 시드 완료")
+            except Exception:
+                db.rollback()
+                raise
+            finally:
+                db.close()
+            return 0
+
+        ensure_tools("uv", "psql")
+        raw_db_url = detect_database_url(args.database_url, prefer="shoppingmall")
+        db_conf = parse_database_url(raw_db_url)
+
+        ensure_postgres_running(db_conf)
+        ensure_database_exists(db_conf)
+
+        initialized = args.mode == "init"
+        if args.mode == "ensure":
+            if args.rebuild_schema:
+                info("사용자 요청으로 강제 초기화 수행 (--rebuild-schema)")
+                uv_sync_backend(args.skip_sync)
+                run_seed_pipeline(raw_db_url, rebuild_schema=True)
+                run_review_seed_pipeline(raw_db_url)
+                initialized = True
+            elif is_shoppingmall_ready(db_conf):
+                info("ShoppingMall DB 상태 정상 (초기화 생략)")
+            else:
+                info("ShoppingMall DB 상태 불완전 (초기화 수행)")
+                uv_sync_backend(args.skip_sync)
+                rebuild_schema = not all_shop_tables_exist(db_conf)
+                if rebuild_schema:
+                    info("shop_* 테이블 일부 누락 감지 (스키마 재생성 모드)")
+                else:
+                    truncate_shop_tables(db_conf)
+                run_seed_pipeline(raw_db_url, rebuild_schema=rebuild_schema)
+                run_review_seed_pipeline(raw_db_url)
+                initialized = True
+        else:
+            uv_sync_backend(args.skip_sync)
+            rebuild_schema = args.rebuild_schema or (not all_shop_tables_exist(db_conf))
+            if rebuild_schema:
+                info("shop_* 테이블 일부 누락 또는 사용자 요청으로 스키마 재생성 모드")
+            else:
+                truncate_shop_tables(db_conf)
+            run_seed_pipeline(raw_db_url, rebuild_schema=rebuild_schema)
+            run_review_seed_pipeline(raw_db_url)
+
+        if initialized:
+            print_db_summary(db_conf, args.verbose_table_info)
+            print()
+            info("ShoppingMall 데이터베이스 초기화 완료")
+        else:
+            info("ShoppingMall 데이터베이스 상태 확인 완료")
         return 0
     except Exception as exc:
-        db.rollback()
         error(str(exc))
         return 1
-    finally:
-        db.close()
 
 
 if __name__ == "__main__":
